requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
fire==0.7.0
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
matplotlib==3.10.1
MouseInfo==0.1.3
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==11.1.0
playsound==1.2.2
protobuf==3.20.2
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytweening==1.2.0
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
